+++ parler/utils/compat.py
@@ -1,11 +1,13 @@
 """
 Django compatibility features
 """
-from django.db import transaction
+from django.db import transaction, models
 
 __all__ = (
     'transaction_atomic',
     'add_preserved_filters',
+    'with_metaclass',
+    'HideChoicesCharField',
 )
 
 # New transaction support in Django 1.6
@@ -37,3 +39,26 @@ def with_metaclass(meta, *bases):
                 return type.__new__(cls, name, (), d)
             return meta(name, bases, d)
     return metaclass('temporary_class', None, {})
+
+
+class HideChoicesCharField(models.CharField):
+    # For Django 1.7, hide the 'choices' for a field.
+    def deconstruct(self):
+        name, path, args, kwargs = models.CharField.deconstruct(self)
+
+        # Hide the fact this model was used.
+        if path == __name__ + '.HideChoicesCharField':
+            path = 'django.db.models.CharField'
+        try:
+            del kwargs['choices']
+        except KeyError:
+            pass
+
+        return name, path, args, kwargs
+
+try:
+    from south.modelsinspector import add_introspection_rules
+except ImportError:
+    pass
+else:
+    add_introspection_rules([], ["^" + __name__.replace(".", "\.") + "\.HideChoicesCharField"])